move withGradings() from app/api/answers/get_current.go into app/api/answers/answers.go
